+++ src/tests/conduit/t_conduit_schema.cpp
@@ -325,8 +325,6 @@ TEST(schema_basics, rename_child)
 
     s.rename_child("c","d");
 
-    s.print();
-
     cnames = s.child_names();
     EXPECT_TRUE(s.has_child("d"));
     EXPECT_FALSE(s.has_child("c"));
@@ -334,7 +332,67 @@ TEST(schema_basics, rename_child)
 }
 
 
+//-----------------------------------------------------------------------------
+TEST(schema_basics, pathlike_child_names)
+{
+    Schema s;
+    
+    std::string shared_name = "a/b";
+    std::string path_only = "c/d";
+    std::string direct_only = "e/f";
 
+    s[shared_name].set(DataType::int64());
+    s.add_child(shared_name).set(DataType::int64());
+
+    s[path_only].set(DataType::int64());
+    s.add_child(direct_only).set(DataType::int64());
+    
+    s.print();
+
+    EXPECT_TRUE(s.has_child(shared_name));
+    EXPECT_TRUE(s.has_path(shared_name));
+
+    EXPECT_TRUE(s.has_path(path_only));
+    EXPECT_FALSE(s.has_path(direct_only));
+
+    EXPECT_TRUE(s.has_child(direct_only));
+    EXPECT_FALSE(s.has_child(path_only));
+
+    // Test that explicitly removing children doesn't remove
+    // by path and vice-versa 
+    std::string second_shared_name = "foo/bar";
+    s[second_shared_name].set(DataType::int64());
+    s.add_child(second_shared_name).set(DataType::int64());
+
+    s.remove(shared_name);
+    s.remove_child(second_shared_name);
+
+    EXPECT_TRUE(s.has_child(shared_name));
+    EXPECT_FALSE(s.has_path(shared_name));
+
+    EXPECT_TRUE(s.has_path(second_shared_name));
+    EXPECT_FALSE(s.has_child(second_shared_name));
+
+    s.print();
+
+    // check compact_to , equal and compatible
+    Schema s2;
+    s["a"].set(DataType::int64());
+    s.add_child("key_with_/_ex").set(DataType::int64());
+    
+    EXPECT_EQ(s.child("key_with_/_ex").path(),"{key_with_/_ex}");
+    
+    Schema s3(s2);
+    
+    EXPECT_TRUE(s2.equals(s3));
+    EXPECT_TRUE(s2.compatible(s3));
+    
+    Schema s2_compact;
+    s2.compact_to(s2_compact);
+    EXPECT_TRUE(s2_compact.is_compact());
+    EXPECT_TRUE(s2.compatible(s2_compact));
+
+}
 
 
 //-----------------------------------------------------------------------------